--- src/teavm/java/com/beanit/asn1bean/ber/types/BerBoolean.java
@@ -1,116 +1,0 @@
-/*
- * Copyright 2012 The ASN1bean Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License
- * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
- * or implied. See the License for the specific language governing permissions and limitations under
- * the License.
- */
-package com.beanit.asn1bean.ber.types;
-
-import com.beanit.asn1bean.ber.BerLength;
-import com.beanit.asn1bean.ber.BerTag;
-import com.beanit.asn1bean.ber.ReverseByteArrayOutputStream;
-import java.io.EOFException;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
-import java.io.Serializable;
-
-public class BerBoolean implements Serializable, BerType {
-
-  public static final BerTag tag =
-      new BerTag(BerTag.UNIVERSAL_CLASS, BerTag.PRIMITIVE, BerTag.BOOLEAN_TAG);
-  private static final long serialVersionUID = 1L;
-  public boolean value;
-  private byte[] code = null;
-
-  public BerBoolean() {}
-
-  public BerBoolean(byte[] code) {
-    this.code = code;
-  }
-
-  public BerBoolean(boolean value) {
-    this.value = value;
-  }
-
-  @Override
-  public int encode(OutputStream reverseOS) throws IOException {
-    return encode(reverseOS, true);
-  }
-
-  public int encode(OutputStream reverseOS, boolean withTag) throws IOException {
-
-    if (code != null) {
-      reverseOS.write(code);
-      if (withTag) {
-        return tag.encode(reverseOS) + code.length;
-      }
-      return code.length;
-    }
-
-    int codeLength = 1;
-
-    if (value) {
-      reverseOS.write(0xff);
-    } else {
-      reverseOS.write(0);
-    }
-
-    codeLength += BerLength.encodeLength(reverseOS, codeLength);
-
-    if (withTag) {
-      codeLength += tag.encode(reverseOS);
-    }
-
-    return codeLength;
-  }
-
-  @Override
-  public int decode(InputStream is) throws IOException {
-    return decode(is, true);
-  }
-
-  public int decode(InputStream is, boolean withTag) throws IOException {
-
-    int codeLength = 0;
-
-    if (withTag) {
-      codeLength += tag.decodeAndCheck(is);
-    }
-
-    BerLength length = new BerLength();
-    codeLength += length.decode(is);
-
-    if (length.val != 1) {
-      throw new IOException("Decoded length of BerBoolean is not correct");
-    }
-
-    int nextByte = is.read();
-    if (nextByte == -1) {
-      throw new EOFException("Unexpected end of input stream.");
-    }
-
-    codeLength++;
-    value = nextByte != 0;
-
-    return codeLength;
-  }
-
-  public void encodeAndSave(int encodingSizeGuess) throws IOException {
-    ReverseByteArrayOutputStream os = new ReverseByteArrayOutputStream(encodingSizeGuess);
-    encode(os, false);
-    code = os.getArray();
-  }
-
-  @Override
-  public String toString() {
-    return "" + value;
-  }
-}
