--- src/teavm/java/com/rhg/rsa/RSABaseInterface.java
@@ -1,15 +1,0 @@
-package com.rhg.rsa;
-
-/**
- * Write a description of interface RSABaseInterface here.
- * 
- * @author Rob
- * @version 06/01/2010
- */
-
-public interface RSABaseInterface
-{
-    byte PUBLIC_KEY = 1;
-    byte PRIVATE_KEY = 2;
-    byte COMPLETE_KEY = 3;
-}
