--- src/teavm/java/com/rhg/rsa/RSAConstants.java
@@ -1,18 +1,0 @@
-package com.rhg.rsa;
-
-import java.math.BigInteger;
-/**
- * Interface providing often used BigInteger constants.
- * 
- * @author Rob
- * @version 05/30/2010
- */
-
-public interface RSAConstants
-{
-    BigInteger ZERO = BigInteger.ZERO;
-    BigInteger ONE = BigInteger.ONE;
-    BigInteger TWO = BigInteger.valueOf(2);
-    BigInteger THREE = BigInteger.valueOf(3);
-    BigInteger TWO_FIFTY_SIX = BigInteger.valueOf(256);
-}
